requirements.txt
# Required for system tray functionality (optional)
pystray>=0.19.0
Pillow>=8.0.0

# Required for global hotkeys (optional)
keyboard>=1.13.0

# Note: All dependencies are optional
# The application will work with just Python standard library
# but with reduced functionality:
# - Without pystray/Pillow: Uses simple window instead of system tray
# - Without keyboard: No global hotkeys